fix: suppress errors when .env does not exist
Hi,
I was planning to add this to the existing version on npm, but when I open the repository, I saw it's already changed. Anyways, in any case, it's good to have this config because otherwise, you might end up with ENOENT errors which is not even related to you and your work.

Thanks in advance!